+++ CHANGELOG.md
@@ -10,6 +10,7 @@ with respect to its command line interface and HTTP interface.
 
 ### Added
 * CLI: If no image is present in runspec, return a fatal flaw in build.
+* Server: Update logging to a more structured format
 
 ### Changed
 * All: error parsing repo from SourceLocation now more informative.